# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickFlow API",
    "description": "Single-user productivity API: tasks, habits, learning resources, plans, dashboard. Errors are RFC 7807 problem+json; validation errors list {field, message} in `errors`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/habits/{id}": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "Get a habit",
        "operationId": "get_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Habits"
        ],
        "summary": "Update a habit",
        "operationId": "update_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Habits"
        ],
        "summary": "Delete a habit and its completions",
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Habit not found"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/tasks/{id}/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Mark a task completed (status DONE)",
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Task not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/habits": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "List habits with progress",
        "description": "active=true/false filters; omit for all",
        "operationId": "list_3",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Habit"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Habit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Create a habit",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HabitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    },
    "/api/habits/{id}/deactivate": {
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Deactivate a habit",
        "operationId": "deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    },
    "/api/habits/{id}/completions": {
      "get": {
        "tags": [
          "Habits"
        ],
        "summary": "List completions, newest first",
        "operationId": "completions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitCompletion"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HabitCompletion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Mark complete for a date (default today); returns the habit with updated progress",
        "operationId": "complete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "409": {
            "description": "Already completed for that date, or habit inactive",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    },
    "/api/habits/{id}/activate": {
      "post": {
        "tags": [
          "Habits"
        ],
        "summary": "Activate a habit",
        "operationId": "activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/habits/{id}/completions/{date}": {
      "delete": {
        "tags": [
          "Habits"
        ],
        "summary": "Undo the completion for a date",
        "operationId": "uncomplete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Habit not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Habit"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "overdue": {
            "type": "boolean",
            "description": "Due before today, not done, not archived"
          }
        }
      },
...
      "HabitRequest": {
        "required": [
          "frequency",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY"
            ]
          }
        }
      },
      "Habit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          },
          "progress": {
            "$ref": "#/components/schemas/HabitProgress"
          }
        }
      },
      "HabitProgress": {
        "type": "object",
        "properties": {
          "completedToday": {
            "type": "boolean"
          },
          "completedThisPeriod": {
            "type": "boolean"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "totalCompletions": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "CompletionRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Defaults to today; future dates are rejected",
            "format": "date"
          }
        }
      },
...
      "HabitCompletion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "habitId": {
            "type": "integer",
            "format": "int64"
          },
          "completionDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}